Move graph-like queries code into is own module
Change-type: patch
